requirements.txt
streamlit~=1.40.2
google-generativeai
python-dotenv~=1.0.1
langchain~=0.3.9
PyPDF2~=3.0.1
chromadb
faiss-cpu
langchain_google_genai
referencing~=0.35.1
Pygments~=2.18.0
pip~=23.3
attrs~=24.2.0
protobuf~=5.29.1
pillow~=11.0.0
filelock~=3.16.1
typing_extensions~=4.12.2
packaging~=24.2
zipp~=3.21.0
importlib_resources~=6.4.5
build~=1.2.2.post1
pyreadline3~=3.5.4
langchain-google-genai~=2.0.6
python-docx~=1.1.2
requests~=2.32.3
beautifulsoup4~=4.12.3